% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{Network centrality}
\usage{
centrality(X)
}
\arguments{
\item{X}{column}
}
\value{
The centrality of the network
}
\description{
Calculate some simple network centrality measures
}
\examples{
temp1 <- cen

centrality(SWen,R1.strength);
}
