% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{getHapaxNodes}
\alias{getHapaxNodes}
\title{Extract Isolated nodes}
\usage{
getHapaxNodes(G, mode)
}
\arguments{
\item{mode}{}
}
\value{
A list of hapax legomena nodes
}
\description{
Isolated nodes aka hapax legomena
}
