% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{getLargestComponent}
\alias{getLargestComponent}
\title{Extract the largest connected component from a graph}
\usage{
getLargestComponent(G, mode)
}
\arguments{
\item{mode}{The mode (weak or strong)}
}
\value{
The strongest connected component subgraph
}
\description{
Extract the largest connected component from a graph
}
