% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{graphSummary}
\alias{graphSummary}
\title{Show graph summary statistics}
\usage{
graphSummary(G)
}
\arguments{
\item{G}{}
}
\value{
description basic network statistics
}
\description{
Show graph summary statistics
}
\examples{
desc = graphSummary(G)
}
