% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGraph.R
\name{importGraph}
\alias{importGraph}
\title{Import association data as unipartite graph}
\usage{
importGraph(X)
}
\arguments{
\item{X}{}
}
\value{

}
\description{
The graph should have two columns, cue, and response
}
\examples{
G = importGraph(SWen)
}
