% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWen}
\alias{SWen}
\title{SWen}
\format{
A data frame with five variables: \code{cue}, \code{response},
  \code{R1}, \code{N} and \code{R1.Strength}.
}
\source{
\url{https://smallworldofwords.org/project/research}
}
\usage{
SWen
}
\description{
Small World of Words English (2018)
}
\keyword{datasets}
