% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USF}
\alias{USF}
\title{USF}
\format{
A data frame with five variables: \code{cue}, \code{response},
  \code{R1}, \code{N} and \code{R1.Strength}.
  Note that the original data have been adapted to balance the number of
  responses per cue, and words were normalized (spelling, capitalization)
}
\source{
\url{https://smallworldofwords.org/project/research}
}
\usage{
USF
}
\description{
University of Florida Word Association Norms
}
\keyword{datasets}
